requirements.txt
pandas
numpy
matplotlib
seaborn
statsmodels
scikit-learn
yfinance
fredapi
plotly
pmdarima
python-dotenv